Added custom flow layout to OpenWith.
Issue #93
